+++ source/Sylvan.Data.Excel/Xlsx/XlsxWorkbookReader.cs
@@ -844,8 +844,7 @@ sealed class XlsxWorkbookReader : ExcelDataReader
 						reader.Read();
 						if (reader.NodeType == XmlNodeType.Text || reader.NodeType == XmlNodeType.SignificantWhitespace)
 						{
-
-							s = reader.Value;
+							s = OpenXmlCodec.DecodeString(reader.Value);
 						}
 						else if (reader.NodeType == XmlNodeType.EndElement && reader.LocalName == "t")
 						{
@@ -858,6 +857,7 @@ sealed class XlsxWorkbookReader : ExcelDataReader
 					}
 					if (c == 0)
 					{
+
 						str = s;
 					}
 					else